# internal/controller/handler/v1/addresses/gen/type.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// AddressCandidate 郵便番号 lookup による住所候補 1 件です。都道府県名は prefectures マスタで解決し、
// 解決できた場合のみ prefectureId を埋めます（解決不能時は null で、市区町村・町域の部分補完は生かします）。
type AddressCandidate struct {
	// City 市区町村
	//
	// Example: 千代田区
	City string `json:"city"`

	// PrefectureId 都道府県ID。県名を prefectures マスタで完全一致解決できなかった場合は null。
	//
	// Example: 101caa1e-84e7-4ceb-9108-50d40b6be1a3
	PrefectureId *openapi_types.UUID `json:"prefectureId"`

	// PrefectureName 都道府県名（外部 lookup が返すフル表記、例 "東京都"）
	//
	// Example: 東京都
	PrefectureName string `json:"prefectureName"`

	// Town 町域
	//
	// Example: 千代田
	Town string `json:"town"`
}

// AddressCandidatesResponse 郵便番号 lookup による住所候補のレスポンスです。
// 外部 lookup サービスの障害時は 503 ではなく `candidates: []` + `isFallback: true` で degrade し、
// 登録フローを止めません（外部が落ちても縦一本のジャーニーを進めるためのフォールバック）。
type AddressCandidatesResponse struct {
	// Candidates 住所候補の一覧。該当なし・degrade 時は空配列。
	Candidates []AddressCandidate `json:"candidates"`

	// IsFallback lookup 機構が機能しなかった（外部不通・不正応答）場合に true。true のときフロントは全項目手入力へ倒します。
	// 正常応答で候補が 0 件（該当なし）の場合や、候補はあるが県名解決に失敗した場合は false です。
	//
	//
	// Example: false
	IsFallback bool `json:"isFallback"`
}

// ErrorResponse エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type ErrorResponse struct {
	// Code 機械的に処理可能なエラーコード
	//
	// Example: EXAMPLE_ERROR
	Code string `json:"code"`

	// Message 人間が読めるエラーメッセージ
	//
	// Example: 定義されたエラーメッセージ
	Message string `json:"message"`

	// RequestId リクエストID
	//
	// Example: abcdefghijklmnopqrstuvwxyz
	RequestId string `json:"requestId"`
}

// PostalCodeParam Example: 100-0001
type PostalCodeParam = string

// BadRequest400 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type BadRequest400 = ErrorResponse

// InternalServerError500 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type InternalServerError500 = ErrorResponse

// MethodNotAllowed405 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type MethodNotAllowed405 = ErrorResponse

// ServiceUnavailable503 エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
type ServiceUnavailable503 = ErrorResponse

// GetAddressesParams defines parameters for GetAddresses.
type GetAddressesParams struct {
	// PostalCode 検索対象の郵便番号。ハイフンあり（`123-4567`）・なし（`1234567`）の両方を受け付けます。
	// ハイフン除去の正規化は handler が行い、内層（usecase / gateway）へは 7 桁の数字列を渡します。
	PostalCode PostalCodeParam `form:"postalCode" json:"postalCode"`
}