# internal/controller/handler/v1/carts/items/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// DeleteCartsMeItem 自分のカートからの明細の削除
	// (DELETE /v1/carts/me/items/{productId})
	DeleteCartsMeItem(ctx *echo.Context, productId ProductIdParam, params DeleteCartsMeItemParams) error
	// PutCartsMeItem 自分のカートへの明細の設定（数量の upsert）
	// (PUT /v1/carts/me/items/{productId})
	PutCartsMeItem(ctx *echo.Context, productId ProductIdParam, params PutCartsMeItemParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// DeleteCartsMeItem converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteCartsMeItem(ctx *echo.Context) error {
	var err error
	// ------------- Path parameter "productId" -------------
	var productId ProductIdParam

	err = runtime.BindStyledParameterWithOptions("simple", "productId", ctx.Param("productId"), &productId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter productId: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params DeleteCartsMeItemParams

	headers := ctx.Request().Header
	// ------------- Optional header parameter "X-Cart-Session" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Cart-Session")]; found {
		var XCartSession CartSessionParam
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for X-Cart-Session, got %d", n))
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Cart-Session", valueList[0], &XCartSession, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false, Type: "string", Format: ""})
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter X-Cart-Session: %s", err))
		}

		params.XCartSession = &XCartSession
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteCartsMeItem(ctx, productId, params)
	return err
}

// PutCartsMeItem converts echo context to params.
func (w *ServerInterfaceWrapper) PutCartsMeItem(ctx *echo.Context) error {
	var err error
	// ------------- Path parameter "productId" -------------
	var productId ProductIdParam

	err = runtime.BindStyledParameterWithOptions("simple", "productId", ctx.Param("productId"), &productId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter productId: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params PutCartsMeItemParams

	headers := ctx.Request().Header
	// ------------- Optional header parameter "X-Cart-Session" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Cart-Session")]; found {
		var XCartSession CartSessionParam
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for X-Cart-Session, got %d", n))
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Cart-Session", valueList[0], &XCartSession, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false, Type: "string", Format: ""})
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter X-Cart-Session: %s", err))
		}

		params.XCartSession = &XCartSession
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutCartsMeItem(ctx, productId, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.DELETE(options.BaseURL+"/v1/carts/me/items/:productId", wrapper.DeleteCartsMeItem, options.OperationMiddlewares["DeleteCartsMeItem"]...)
	router.PUT(options.BaseURL+"/v1/carts/me/items/:productId", wrapper.PutCartsMeItem, options.OperationMiddlewares["PutCartsMeItem"]...)

}

type BadRequest400JSONResponse ErrorResponse

type Conflict409JSONResponse ErrorResponse

type InternalServerError500JSONResponse ErrorResponse

type MethodNotAllowed405JSONResponse ErrorResponse

type ServiceUnavailable503JSONResponse ErrorResponse

type Unauthorized401JSONResponse ErrorResponse

type UnprocessableEntity422JSONResponse ErrorResponseWithDetails

type DeleteCartsMeItemRequestObject struct {
	ProductId ProductIdParam `json:"productId"`
	Params    DeleteCartsMeItemParams
}

type DeleteCartsMeItemResponseObject interface {
	VisitDeleteCartsMeItemResponse(w http.ResponseWriter) error
}

type DeleteCartsMeItem204Response struct {
}

func (response DeleteCartsMeItem204Response) VisitDeleteCartsMeItemResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeleteCartsMeItem400JSONResponse struct{ BadRequest400JSONResponse }

func (response DeleteCartsMeItem400JSONResponse) VisitDeleteCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteCartsMeItem401JSONResponse struct{ Unauthorized401JSONResponse }

func (response DeleteCartsMeItem401JSONResponse) VisitDeleteCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteCartsMeItem405JSONResponse struct {
	MethodNotAllowed405JSONResponse
}

func (response DeleteCartsMeItem405JSONResponse) VisitDeleteCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(405)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteCartsMeItem500JSONResponse struct {
	InternalServerError500JSONResponse
}

func (response DeleteCartsMeItem500JSONResponse) VisitDeleteCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type DeleteCartsMeItem503JSONResponse struct {
	ServiceUnavailable503JSONResponse
}

func (response DeleteCartsMeItem503JSONResponse) VisitDeleteCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

type PutCartsMeItemRequestObject struct {
	ProductId ProductIdParam `json:"productId"`
	Params    PutCartsMeItemParams
	Body      *PutCartsMeItemJSONRequestBody
}

type PutCartsMeItemResponseObject interface {
	VisitPutCartsMeItemResponse(w http.ResponseWriter) error
}

type PutCartsMeItem200JSONResponse CartResponse

func (response PutCartsMeItem200JSONResponse) VisitPutCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type PutCartsMeItem400JSONResponse struct{ BadRequest400JSONResponse }

func (response PutCartsMeItem400JSONResponse) VisitPutCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PutCartsMeItem401JSONResponse struct{ Unauthorized401JSONResponse }

func (response PutCartsMeItem401JSONResponse) VisitPutCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type PutCartsMeItem405JSONResponse struct {
	MethodNotAllowed405JSONResponse
}

func (response PutCartsMeItem405JSONResponse) VisitPutCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(405)
	_, err := buf.WriteTo(w)
	return err
}

type PutCartsMeItem409JSONResponse struct{ Conflict409JSONResponse }

func (response PutCartsMeItem409JSONResponse) VisitPutCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type PutCartsMeItem422JSONResponse struct {
	UnprocessableEntity422JSONResponse
}

func (response PutCartsMeItem422JSONResponse) VisitPutCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)
	_, err := buf.WriteTo(w)
	return err
}

type PutCartsMeItem500JSONResponse struct {
	InternalServerError500JSONResponse
}

func (response PutCartsMeItem500JSONResponse) VisitPutCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type PutCartsMeItem503JSONResponse struct {
	ServiceUnavailable503JSONResponse
}

func (response PutCartsMeItem503JSONResponse) VisitPutCartsMeItemResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// DeleteCartsMeItem 自分のカートからの明細の削除
	// (DELETE /v1/carts/me/items/{productId})
	DeleteCartsMeItem(ctx context.Context, request DeleteCartsMeItemRequestObject) (DeleteCartsMeItemResponseObject, error)
	// PutCartsMeItem 自分のカートへの明細の設定（数量の upsert）
	// (PUT /v1/carts/me/items/{productId})
	PutCartsMeItem(ctx context.Context, request PutCartsMeItemRequestObject) (PutCartsMeItemResponseObject, error)
}

type StrictHandlerFunc func(ctx *echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// DeleteCartsMeItem operation middleware
func (sh *strictHandler) DeleteCartsMeItem(ctx *echo.Context, productId ProductIdParam, params DeleteCartsMeItemParams) error {
	var request DeleteCartsMeItemRequestObject

	request.ProductId = productId
	request.Params = params

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteCartsMeItem(ctx.Request().Context(), request.(DeleteCartsMeItemRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteCartsMeItem")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(DeleteCartsMeItemResponseObject); ok {
		return validResponse.VisitDeleteCartsMeItemResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PutCartsMeItem operation middleware
func (sh *strictHandler) PutCartsMeItem(ctx *echo.Context, productId ProductIdParam, params PutCartsMeItemParams) error {
	var request PutCartsMeItemRequestObject

	request.ProductId = productId
	request.Params = params

	var body PutCartsMeItemJSONRequestBody
	var err error
	if _, ok := ctx.Echo().Binder.(*echo.DefaultBinder); ok {
		// Bind only the request body, so that path and query parameters
		// are not also bound into the body struct.
		err = echo.BindBody(ctx, &body)
	} else {
		// A custom binder is installed on the Echo instance; defer to it
		// entirely, since echo.Binder does not expose body-only binding.
		err = ctx.Bind(&body)
	}
	if err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PutCartsMeItem(ctx.Request().Context(), request.(PutCartsMeItemRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PutCartsMeItem")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PutCartsMeItemResponseObject); ok {
		return validResponse.VisitPutCartsMeItemResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}