# openapi/openapi.gen.yaml
openapi: 3.0.3
info:
  title: Go Boilerplate API
  termsOfService: https://github.com/Tomy-ch/go-boilerplate-echo-oapi-sqlc
  contact:
    name: TomyC258460
    url: https://x.com/TomyC258460
  version: 2.2.0
  description: |
    Reference specification for the sample API (users, products, purchases). It exercises every layer
    of the Onion Architecture end to end: cursor and offset pagination, creation and partial update,
    soft delete, idempotency keys, and role-based authorization.

    Handlers (oapi-codegen) and the published reference documentation are both generated from this
    file, so every endpoint change starts here.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
    x-internal: true
  - url: https://api.example.com
    description: Production server (replace with actual URL)
    x-internal: false
  - url: https://staging.api.example.com
    description: Staging server (replace with actual URL)
    x-internal: false
  - url: https://development.api.example.com
    description: Development server (replace with actual URL)
    x-internal: true
security: []
...
paths:
...
  /v1/carts/me/items/{productId}:
    put:
      operationId: PutCartsMeItem
      tags:
        - v1/carts/items
      summary: 自分のカートへの明細の設定（数量の upsert）
      description: |
        呼び出し主体のカートに、指定した商品の数量を設定します。加算ではなく設定なので、
        同じ要求を何回送っても結果は変わりません。**冪等性は明細の自然キー (cart_id, product_id) から
        来るため、Idempotency-Key を必要としません**（自然キーを持たない POST /v1/purchases との対比）。
        主体は認証済みユーザー（Bearer トークンの内部 UserID）か、ゲスト（X-Cart-Session ヘッダ）の
        いずれかで、両方が提示された場合は認証済みユーザーが優先されます。
        認証は任意ですが、**提示された資格情報が無効な場合は匿名として通さず 401 を返します**
        （ADR-0020 (optional-authentication-fail-closed)）。
        **カートがまだ無い主体にはこの操作がカートを作ります**。ゲストの場合はセッショントークンを
        発行して sessionToken に載せて返すので、以降のリクエストで X-Cart-Session に載せてください。
        提示されたトークンでカートを引けなかった場合、その値では作らず新しい値を発行します
        （トークンの秘匿だけがゲストカートへの到達を守っているため、値をクライアントに選ばせません）。
        **有効期限を過ぎたカートは空から始まります**。所有者が確定したカートは行を保ったまま明細を捨て、
        ゲストは新しいカートを作ります。いずれも取得（GET）が期限切れを空のカートとして見せる契約と一致します。
        設定後、応答はカート全体を明細ごとの再評価つきで返します（GET と同じ内容です）。
        在庫は押さえません。在庫不足は投入を拒む理由ではなく、明細の issues として現れます。
        本 op 自体は DB の SELECT / INSERT / UPDATE のみで外部依存を持ちませんが、認証段
        （外部 IdP の JWKS 参照）の一時障害で応答不能となり得るため、認証を伴う op の先例に倣い 503 を宣言します。
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
        - $ref: '#/components/parameters/CartSessionParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemPutRequest'
      responses:
        '200':
          description: 明細の設定に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '405':
          $ref: '#/components/responses/MethodNotAllowed405'
        '409':
          $ref: '#/components/responses/Conflict409'
        '422':
          $ref: '#/components/responses/UnprocessableEntity422'
        '500':
          $ref: '#/components/responses/InternalServerError500'
        '503':
          $ref: '#/components/responses/ServiceUnavailable503'
    delete:
      operationId: DeleteCartsMeItem
      tags:
        - v1/carts/items
      summary: 自分のカートからの明細の削除
      description: |
        呼び出し主体のカートから、指定した商品の明細を取り除きます。
        **対象の明細が無くても成功します**。「無かった」と「消した」を呼び出し側に区別させないため、
        404 を持ちません。主体の決まり方は設定（PUT）と同じで、両方が提示された場合は認証済みユーザーが
        優先されます。認証は任意ですが、**提示された資格情報が無効な場合は匿名として通さず 401 を返します**
        （ADR-0020 (optional-authentication-fail-closed)）。
        **カートを持たない主体が呼んでもカートは作りません**。提示されたセッショントークンでカートを
        引けなかった場合も採番し直さず、何もせずに成功を返します（応答が本文を持たないため、新しい
        トークンを返す場所がありません）。
        **商品の存在も公開状態も確認しません**。非公開になった商品こそカートから外したいので、ここで
        拒むと再評価で unpublished が立った明細を利用者が取り除けなくなります（設定は 422 で拒むのに対し、
        こちらは取り除けます）。
        **明細ごとの再評価を行いません**。この応答は本文を持たず価格を 1 つも提示していないため、提示価格を
        記録すると、次の取得で立つはずの priceIncreased を消してしまいます。
        有効期限は延びます（削除も利用のため）。明細が 0 件になってもカート自体は残ります。
        本 op 自体は DB の SELECT / UPDATE / DELETE のみで外部依存を持ちませんが、認証段
        （外部 IdP の JWKS 参照）の一時障害で応答不能となり得るため、認証を伴う op の先例に倣い 503 を宣言します。
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
        - $ref: '#/components/parameters/CartSessionParam'
      responses:
        '204':
          description: 明細を取り除きました（対象が無かった場合を含みます）
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '405':
          $ref: '#/components/responses/MethodNotAllowed405'
        '500':
          $ref: '#/components/responses/InternalServerError500'
        '503':
          $ref: '#/components/responses/ServiceUnavailable503'
...
components:
  parameters:
...
    CartSessionParam:
      name: X-Cart-Session
      in: header
      description: |
        ゲストカートのセッショントークン。所有者が未確定のカートへ到達する唯一の手段で、
        未認証の呼び出し側はこのヘッダで自分のカートを指定する。
        認証済みの呼び出し側では無視される（カートは内部 UserID から解決されるため）。
        値は 256 ビットを base64url（パディング無し）で表現した 43 文字。
      required: false
      schema:
        type: string
        pattern: ^[A-Za-z0-9_-]{43}$
        minLength: 43
        maxLength: 43
        example: b3JnLWV4YW1wbGUtc2Vzc2lvbi10b2tlbi1zYW1wbGU
...
    ProductIdParam:
      name: productId
      in: path
      required: true
      description: 商品のUUID
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        maxLength: 36
        example: b1d4e0f2-3c5a-4b6d-8e7f-1a2b3c4d5e6f
...
  securitySchemes:
    BearerAuth:
      type: http
      description: |
        認証にBearerトークンを使用することを示します。
        JWT の検証は RFC8725（JSON Web Token Best Current Practices）に従います。
        署名アルゴリズムは許可リストで固定し、`alg: none` と鍵の混同を拒否します。
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      description: 認証にBasic認証を使用することを示します。
      scheme: basic
  schemas:
    ErrorResponse:
      type: object
      description: |
        エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: 機械的に処理可能なエラーコード
          example: EXAMPLE_ERROR
        message:
          type: string
          description: 人間が読めるエラーメッセージ
          example: 定義されたエラーメッセージ
        requestId:
          type: string
          description: リクエストID
          example: abcdefghijklmnopqrstuvwxyz
...
    CartItemIssue:
      type: string
      description: |
        カート明細の再評価結果。取得のたびに商品の現在値と突き合わせて判定される。
        1 つの明細に複数同時に立ちうるが、notFound は単独で立ち（商品が引けない以上、在庫も価格も
        判定材料が無い）、outOfStock と insufficientStock は排他（在庫 0 は「不足」ではなく「無い」）。
        いずれの値もカートの取得を失敗にしない。買えない明細があることは要求の不正ではないため。
      enum:
        - notFound
        - unpublished
        - outOfStock
        - insufficientStock
        - priceIncreased
        - priceDecreased
      example: outOfStock
    CartItemResponse:
      type: object
      description: |
        カート明細 1 件。商品の現在値との突き合わせ結果を issues に添える。
        カートは商品の名前も価格も画像も保持しないため、productName / unitPrice / imagePath は
        取得時点の商品の値であり、購入時の金額を拘束しない（請求額を確定するのは購入明細のスナップショット）。
      required:
        - productId
        - quantity
        - issues
      properties:
        productId:
          type: string
          format: uuid
          description: 商品 ID。カート内で明細を一意に指す自然キー。
          example: 0198a1b2-c3d4-7e5f-8a9b-0c1d2e3f4a5b
        productName:
          type: string
          nullable: true
          description: 商品名。商品を引けなかった場合（issues に notFound）は null です。
          example: エルゴノミクスチェア
        imagePath:
          type: string
          nullable: true
          description: |
            取得時点の商品の代表画像のオブジェクトキー。代表画像とは、出品者が並べた表示順の先頭にある
            画像を指します。表示 URL は呼び出し側が配信ベース URL と組み立てます。
            代表画像を持たない商品と、商品を引けなかった明細（issues に notFound）では null です。
          example: products/b1d4e0f2-3c5a-4b6d-8e7f-1a2b3c4d5e6f.png
        quantity:
          type: integer
          format: int32
          description: カートに入れた数量。
          example: 2
        unitPrice:
          type: string
          pattern: ^\d+(\.\d+)?$
          nullable: true
          description: |
            取得時点の商品の単価。サブセント精度を保持する decimal 文字列（USD ドル）で表します（例 "19.99"）。
            商品を引けなかった場合（issues に notFound）は null です。
            JSON number は IEEE754 double として復元され精度を失うため、文字列で表現します。
          example: '19.99'
        issues:
          type: array
          description: |
            この明細の再評価結果。空配列なら現時点で購入可能です。
            subtotalAmount の合算対象になるのは、この配列が空の明細だけです。
          items:
            $ref: '#/components/schemas/CartItemIssue'
        availableQuantity:
          type: integer
          format: int32
          nullable: true
          description: |
            issues に insufficientStock がある場合の、今買える上限（商品の在庫数）。
            それ以外の場合は null です。
          example: 1
    CartResponse:
      description: |
        カートのレスポンススキーマ。カート参照・投入・削除の各操作で共有します。
        カートは在庫を押さえず、確定単価も持ちません。買うつもりの控えであって、
        売り越しの禁止も請求額の確定も購入成立時の関心です。
      type: object
      required:
        - items
        - subtotalAmount
      properties:
        sessionToken:
          type: string
          nullable: true
          description: |
            ゲストカートのセッショントークン。以降のリクエストで X-Cart-Session ヘッダに載せます。
            所有者が確定済みのカート、および新規発行が起きなかった操作では null です。
            値が入るのは、その操作がゲストカートを新しく作った場合だけです
            （取得はカート行を作らないため、常に null になります）。
          example: b3JnLWV4YW1wbGUtc2Vzc2lvbi10b2tlbi1zYW1wbGU
        items:
          type: array
          description: |
            カートの明細。1 件も入っていない場合、およびカートがまだ無い場合は空配列です。
          items:
            $ref: '#/components/schemas/CartItemResponse'
        subtotalAmount:
          type: integer
          format: int64
          description: |
            小計。issues が空の明細（現時点で購入可能なもの）だけを単価×数量で合算した参考値で、
            USD セント単位の整数です。請求額ではありません。
            合算は decimal のまま行い、最後に一度だけ最小単位へ丸めます。
            合算対象が 1 件も無い場合は 0 です。
          example: 3998
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: |
            カートの有効期限。操作のたびに延長されます。
            カートがまだ無い場合は行そのものが存在しないため null です
            （明細を空にしただけの場合はカート行が残るため null になりません）。
          example: '2026-09-13T00:00:00Z'
...
    ErrorResponseWithDetails:
      type: object
      description: |
        詳細識別子付きエラーレスポンススキーマ。 この schema を error レスポンスに宣言したエンドポイントだけが details を返せる（opt-in）。 未宣言のエンドポイントでは errorhandler が details を fail-closed で落とす。
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: 機械的に処理可能なエラーコード
          example: EXAMPLE_ERROR
        message:
          type: string
          description: 人間が読めるエラーメッセージ
          example: 定義されたエラーメッセージ
        requestId:
          type: string
          description: リクエストID
          example: abcdefghijklmnopqrstuvwxyz
        details:
          type: array
          description: エラー詳細を示す公開可能な識別子(任意)。検証エラーでは不正なフィールド名が入る
          items:
            type: string
          example:
            - firstName
            - email
...
    CartItemPutRequest:
      type: object
      description: |
        カート明細の数量設定リクエスト（application/json）。対象の商品はパスで指定します。
        数量は加算ではなく設定です。同じ要求を 2 回送っても結果は変わりません。
        0 は削除ではなく範囲外として 400 を返します（削除は DELETE が担い、1 つの操作に 2 つの意味を持たせません）。
      additionalProperties: false
      required:
        - quantity
      properties:
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 99
          description: 設定する数量。現在の数量に加算するのではなく、この値で置き換えます。 範囲外の値はドメインに届く前に 400 で拒否されます。
          example: 2
...
  responses:
    MethodNotAllowed405:
      description: 許可されていないリクエスト方法です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest400:
      description: リクエストが不正です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized401:
      description: 認証が必要です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError500:
      description: サーバーエラーが発生しました。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable503:
      description: サービスが一時的に利用できません。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict409:
      description: リソースが競合しています。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity422:
      description: 入力値が要件を満たしていません。details に不正なフィールド名が入ります。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWithDetails'
...